requirements.txt
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
azure-functions>=1.17.0
playwright>=1.40.0
